pyproject.toml
[tool.black]
line-length = 180

[tool.flake8]
max-line-length = 172
ignore = "E203"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "C",  # flake8-comprehensions
]
line-length = 172
# Allow unused variables when underscore-prefixed.
ignore = [
    "E501",
    "C408", # Unnecessary dict call - rewrite as a literal.
    "C901", # Function is too complex
    "C416", # Unnecessary list comprehension
]

[tool.poetry]
name = "pmj"
version = "0.1.0"
description = "Pimp my jira"
authors = ["Manuel Bovo <manuel.bovo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
PyYAML = "^6.0"
colorama = "^0.4.4"
rich = "^13.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.5.0"
pytest-flake8 = "^1.0.7"
pytest-mock = "^3.6.1"
ruff = "^0.1.11"

[tool.poetry.scripts]
pmj = "pmj.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

